Editor scrolling fix, PyDOS updates
